requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.1
playwright>=1.40
PyMuPDF>=1.24
pytesseract>=0.3.10
pdf2image>=1.17
opencv-python-headless>=4.9
Pillow>=10.0
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
python-dotenv>=1.0
pydantic>=2.5
pydantic-settings>=2.1
rich>=13.7
click>=8.1
tenacity>=8.2
megadl>=0.1
openai>=1.0
jinja2>=3.1
flask>=3.0
gunicorn>=22.0
